use model struct for base health summary
